feat: Add media ID to Media struct
This commit adds a new field, `ID`, to the `Media` struct. The `ID` is used as a unique identifier for each media item and replaces the previous use of `TmdbID`. Additionally, this commit updates all relevant functions to use the new `ID` field instead of `TmdbID`. Finally, this commit also includes minor refactoring and code cleanup.