feat: Add scannerMutex to prevent concurrent scanning
This commit adds a new mutex lock called scannerMutex to prevent concurrent scanning of movie and TV folders. The scanLock field in the MovieScanner and TVScanner structs have been removed and replaced with the new scannerMutex. The TryLock method is used to check if the scanner is currently running before starting a new scan.