feat: Add uploadLock to prevent concurrent uploads and scans
This commit adds a new feature to prevent concurrent uploads and scans. The code changes include adding a new sync.Mutex variable called uploadLock, which is used to lock the upload process when it's already in progress. If an attempt is made to start another upload or scan while one is already running, an error message will be returned. This feature has been added to both the ScanMovies() and ScanTV() functions as well as the UploadMovie() and UploadTV() functions.