feat: Update media-go-pkg to v0.5.2
This commit updates the dependency on media-go-pkg from v0.5.1 to v0.5.2 in go.mod and go.sum files, which includes bug fixes and performance improvements.

It also adds a new "Name" field to the Media struct in root.go, which is used for indexing movies and TV episodes with their names instead of just IDs.

Finally, it refactors the clearDuplicatedMovie() and clearDuplicatedEpisode() functions in root.go by removing the unused fileDestination parameter and renaming some variables for clarity.