feat: Add endpoints to get job logs and job name
This commit adds two new endpoints to the scan controller that allow users to retrieve the logs of the last/current job and the name of the last/current job. The `/scan/logs` endpoint returns an array of `jobLogResponse`, while the `/scan/job-name` endpoint returns a string. Both endpoints have been added to `InitScanController`.