feat: Update media-go-pkg to v1.5.3
The media-indexer module now requires version 1.5.3 of the media-go-pkg package instead of v1.4.3, which was previously required. This change is reflected in both go.mod and go.sum files, with no other significant changes made to either file except for an indirect dependency update to aws-sdk-go v1.44.287 in go.mod file.