feat: Add object storage functionality to media scanner
This commit adds the ability to upload media files to an object storage service. The `MovieScanner` and `TVScanner` structs now have an additional field for an `ObjectStorage` object, which is used to upload the media files after they are processed. The uploaded files are stored in S3 under a directory named after their respective movie or TV show ID. After uploading, the local copy of the file is removed.