feat: Add support for 219 intro file path
This commit adds support for a new intro file path to the media repository. The `NewMediaRepository` function now takes an additional argument, `intro219FilePath`, which is used in the transcoding process. The `IndexMovie` and `IndexTvEpisode` functions have been updated to use this new argument when calling the transcoder. Additionally, the `retrieveMovieList` and `retrieveTvList` functions in the `scan.go` file have been updated to limit the number of goroutines using a semaphore channel.